Update contract and unit tests
